Crate libra_secure_json_rpc

The purpose of the JsonRpcClient presented here is to provide a lightweight and secure JSON RPC client to talk to the JSON RPC service offered by Libra Full Nodes. This is useful for various security-critical components (e.g., the secure key manager), as it allows interaction with the Libra blockchain in a minimal and secure manner.

Note: While a JSON RPC client implementation already exists in the Libra codebase, this provides a simpler and (hopefully) more secure implementation with fewer dependencies.

Structs

Bytes
Represents a vector of bytes using hex encoding.

JsonRpcClient
Provides a lightweight JsonRpcClient implementation.

Enums

Error